status-tracker/requirements.txt
# ── Core runtime ──────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3

# ── Environment loading ───────────────────────────────────────────────────────
# python-dotenv is NOT optional — load_dotenv() in main.py depends on it.
# Without this package installed, the .env file is silently ignored entirely.
python-dotenv==1.0.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.4
httpx==0.28.1
pytest-asyncio==0.24.0
